fixed the windows stuff

+++ src/chunk.cc
@@ -11,8 +11,8 @@ ChunkHeader chunk_header(int id)
 
 bool write_chunk_header(const ChunkHeader *hdr, goat3d_io *io)
 {
-	io->seek(-hdr->size, SEEK_CUR, io->cls);
-	if(io->write(hdr, sizeof *hdr, io->cls) < (ssize_t)sizeof *hdr) {
+	io->seek(-(long)hdr->size, SEEK_CUR, io->cls);
+	if(io->write(hdr, sizeof *hdr, io->cls) < (long)sizeof *hdr) {
 		return false;
 	}
 	return true;
@@ -20,7 +20,7 @@ bool write_chunk_header(const ChunkHeader *hdr, goat3d_io *io)
 
 bool read_chunk_header(ChunkHeader *hdr, goat3d_io *io)
 {
-	if(io->read(hdr, sizeof *hdr, io->cls) < (ssize_t)sizeof *hdr) {
+	if(io->read(hdr, sizeof *hdr, io->cls) < (long)sizeof *hdr) {
 		return false;
 	}
 	return true;
